feat: migrate AU-02 invite signup/activation, delete features-v1/auth (#56)
features-v1/auth/InviteSignupScreen.tsx를 features/auth/InviteScreen.tsx로
이식했다. AU-01·AU-03과 같은 패턴 — v1 구현이 v2 케이스 계약(14종, 변형
A/B, 동의 항목 개수·문구)과 정확히 일치해 새로 짜지 않고 옮겼다.

공용 인프라(inviteApi·inviteTypes·consents)는 v1에서 복사했다.
inviteStates.ts는 resolveInviteState + InviteState 타입만 옮겼다 —
checkPasswordPolicy는 AU-03(#54)에서 이미 features/auth/passwordPolicy.ts로
분리했으므로 다시 만들지 않는다. components/ProtectedRoute.tsx는 v1
안에서도 아무 데서도 쓰이지 않는 고아 파일이라 가져오지 않았다.

이걸로 AU 3장(AU-01·02·03) 이식이 끝나서 features-v1/auth를 통째로
삭제했다 — 더는 그 폴더를 참조하는 v1 화면이 없다(curriculum·dashboard·
trainees는 이 폴더와 무관하며 typecheck로 확인함). main.tsx의 이제는
낡은 "features-v1/auth가 아니라" 주석도 정리했다.

closes #56
